Split lib.rs into multiple modules
